Repair the map function
The map function will now work correctly without panicking so queries
like this work correctly:

    from(db: "telegraf")
        |> range(start: -5m)
        |> filter(fn: (r) => r._measurement == "cpu" and r._field == "usage_user")
        |> group(by: ["_measurement"])
        |> mean()
        |> map(fn: (r) => {mean: r._value})

In this query, the map didn't work correctly because the object
expression did not update its reported set of types. It reported an
object type with the `mean` being an unknown type rather than evaluating
the type based on the variable declarations. This is now fixed.

This also fixes the implementation of map so it doesn't use a type
assertion that panics. It creates a record and then iterates over the
result to add it to that record so a map always returns an object type.